+++ src/main/kotlin/io/openfuture/chain/domain/transaction/data/BaseTransactionData.kt
@@ -3,7 +3,9 @@ package io.openfuture.chain.domain.transaction.data
 abstract class BaseTransactionData(
     val amount: Long,
     val recipientKey: String,
+    val recipientAddress: String,
     val senderKey: String,
+    val senderAddress: String,
     val senderSignature: String
 ) {