+++ src/main/kotlin/io/openfuture/chain/repository/Repositories.kt
@@ -3,6 +3,7 @@ package io.openfuture.chain.repository
 import io.openfuture.chain.entity.SeedWord
 import io.openfuture.chain.entity.Stakeholder
 import io.openfuture.chain.entity.Block
+import io.openfuture.chain.entity.Wallet
 import io.openfuture.chain.entity.peer.Delegate
 import io.openfuture.chain.entity.peer.Peer
 import io.openfuture.chain.entity.transaction.BaseTransaction
@@ -64,5 +65,12 @@ interface DelegateRepository : PeerRepository<Delegate>
 interface StakeholderRepository : BaseRepository<Stakeholder> {
 
     fun findOneByPublicKey(publicKey: String): Stakeholder?
+
+}
+
+@Repository
+interface WalletRepository : BaseRepository<Wallet> {
+
+    fun findOneByAddress(address: String): Wallet?
 
 }