+++ src/main/kotlin/io/openfuture/chain/service/DefaultBlockService.kt
@@ -8,20 +8,32 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class DefaultBlockService(
-    private val blockRepository: BlockRepository
+    private val repository: BlockRepository,
+    private val wallerService: WalletService
 ) : BlockService {
 
     @Transactional(readOnly = true)
-    override fun get(hash: String): Block = blockRepository.findByHash(hash)
+    override fun get(hash: String): Block = repository.findByHash(hash)
         ?: throw NotFoundException("Block with hash:$hash not found")
 
     @Transactional(readOnly = true)
-    override fun getLast(): Block = blockRepository.findFirstByOrderByHeightDesc()
+    override fun getLast(): Block = repository.findFirstByOrderByHeightDesc()
         ?: throw NotFoundException("Last block not found!")
 
     @Transactional(readOnly = true)
     override fun getLastGenesis(): Block {
         TODO("not implemented")
+    }
+
+    @Transactional
+    override fun save(block: Block): Block {
+        val savedBlock = repository.save(block)
+
+        savedBlock.transactions.forEach {
+            wallerService.updateByTransaction(it)
+        }
+
+        return savedBlock
     }
 
 }