+++ src/main/kotlin/io/openfuture/chain/service/DefaultWalletService.kt
@@ -1,0 +1,35 @@
+package io.openfuture.chain.service
+
+import io.openfuture.chain.entity.Wallet
+import io.openfuture.chain.entity.transaction.BaseTransaction
+import io.openfuture.chain.repository.WalletRepository
+import org.springframework.stereotype.Service
+
+@Service
+class DefaultWalletService(
+    private val repository: WalletRepository
+) : WalletService {
+
+    companion object {
+        private const val DEFAULT_WALLET_BALANCE = 0.0
+    }
+
+
+    override fun getBalance(address: String): Double =
+        repository.findOneByAddress(address)?.balance ?: DEFAULT_WALLET_BALANCE
+
+    override fun updateByTransaction(transaction: BaseTransaction) {
+        updateByAddress(transaction.senderAddress, -transaction.amount)
+
+        updateByAddress(transaction.recipientAddress, transaction.amount)
+    }
+
+    private fun updateByAddress(address: String, amount: Long) {
+        val wallet = repository.findOneByAddress(address) ?: Wallet(address)
+
+        wallet.balance += amount
+
+        repository.save(wallet)
+    }
+
+}